fix(instrument-tuner): use new tuning system in routes.tsx
The old pages/tuner.tsx was still being used because routes.tsx
imported from it instead of using the new routes/index.tsx.

- Delete old pages/tuner.tsx
- Update routes.tsx with full new TunerPage implementation
- Includes all new features: multi-instrument tunings, URL sharing,
  custom tuning builder
